Clear the downloading history when test/ directory is empty
Some people stick to working in one directory. We don't recommend it because we recommend to add tests manually and/or to generate tests to increase their ratings, but this is less understood. When test/ directory is empty, It seems we should show some warnings to recommend using separate directories, or just clear the history.